Fix test for HBP in combined network
This has to be done to avoid the first layer to access attributes
that are not initialized since the backward hook might not be called
because there are no further leaf variables after this layer.

So far, it still works because of a second hook that is installed
on top (and had leaf variables, so it was also called) in the
combined layer.

As I am planning to remove this redundant hook in the future to
save memory, the test will raise an exception when running the tests.

Took some time to understand this behavior...