[REF] Improve batch axis detection (#208)
Sub-sampling relies on slicing tensors along their batch axis. So far we detect
the batch axis based on the module. However, BackPACK's custom `Permute`
module can alter the axes of its input tensor and therefore modify the batch axis
position. To account for this, the batch axis not only has to be determined based
on the module, but also on the direction of the tensor, i.e. `input` or `output`.

This PR introduces such a tensor-agnostic batch size detection. Progress on #12:
- Replace hard-coded batch axes in tests with dynamical ones
- Make `jac_t_mat_prod` tests work for `Permute` module
- Support arbitrary batch axes in `subsample`

---

* [ADD] Make batch_axis depend on IO tensor

The `Permute` module can modify the batch axis position in its forward
pass. Therefore, input and output tensors have different batch axes.

* [FIX] Make `jac_t_mat_prod` tests work for `Permute`

* [REF] Replace hard-coded batch axis in test case forward pass

* [DEL] Remove equality check of batch_axis_in/out

* [REF] Generalize slicing in `subsample` to arbitrary axes