[DOC] Wrong IHVP if parameter occurs multiple times in grad
Evidence that the method proposed in #213 does not work whenever a
parameter is used in the first `backward` pass to compute the gradient.
The IHVP is missing terms from the parameter's operation that is used in
the backward pass, because `BatchGrad` only installs the Jacobians of
modules from the forward pass.