[FIX] Make test implementation of GGN work with batch size = 1
The old implementation only rescales the GGN if the loss reduction is
`'mean'` (detectable only if `batch_size > 1`). This misses the case
where `batch_size = 1`. If we abuse `subsampling = [0, 0]` to re-use a
sample multiple times, the GGN won't be rescaled, although a
multiplication by `2` would be required. This is now fixed.