[ADD] `SqrtGGN{Exact,MC} support for ResNet modules (#246)
Support includes `SumModule`, `ScaleModule`, and `Identity`.
Specify accumulation of backpropagated `SqrtGGN{Exact,MC}` info.
Fix incorrect detection of reduction factor in the tests if `batch_size=1`.

---

* [FIX] Make test implementation of GGN work with batch size = 1

The old implementation only rescales the GGN if the loss reduction is
`'mean'` (detectable only if `batch_size > 1`). This misses the case
where `batch_size = 1`. If we abuse `subsampling = [0, 0]` to re-use a
sample multiple times, the GGN won't be rescaled, although a
multiplication by `2` would be required. This is now fixed.

* [ADD] Support ResNet modules in `SqrtGGN{Exact, MC}` extension

* [DEL] Remove print statement

Co-authored-by: Felix Dangel <fdangel@tue.mpg.de>