[DOC] Fully document `HBPConv2d` (#290)
Add documentation to the `HBP` extension of `Conv2d` layers, which is 
responsible to compute `KFAC/KFLR/KFRA`. The docstrings draw connections
to the notation in the [KFC paper](https://arxiv.org/pdf/1602.01407.pdf), and 
outline important differences, as well as improvements for consistency. 
Also add a test case for `KFAC, KFLR` for which both approximations become exact.

Note to myself: I made notes how to connect Hessian backpropagation to `KFAC` 
for convolutions by imposing a Kronecker structure on the backpropagated quantity.
This concept can also be applied to `KFRA` to achieve more consistency, but is 
currently not done by the code.

* [DOC] Fully document `HBPConv2d`

* [TEST] KFAC/KFLR for convolution with single output

Convolution layers with a single output behave like linear layers, as
the weights are not shared over the input.

* [DOC] Polish docstrings

* [TEST] Add integration test for KFRA